refactor: extract codex app-server stdio transport into its own file (#661)
Pure behavior-preserving move: the JSON-RPC send/request/notify path, the
single long-lived stdout reader lifecycle (startStdoutReader /
recoverReaderPanic / scanTerminalError), the per-read timeout machinery
(readLine / readLineOnce / appServerReadTimeoutError), and the protocol
line helpers move from codex_app_server.go (877 lines) to
codex_app_server_transport.go, leaving the runner entry point at 672
lines — under the 800-line production-file budget. The file's entry is
removed from the file-size baseline so the ratchet now enforces the
budget for it directly.
